Share code between standard and fp implementation
