+++ package.json
@@ -2,7 +2,8 @@
   "name": "referer-parser",
   "version": "0.0.4",
   "description": "library for extracting marketing attribution data (such as search terms) from referrer URLs",
-  "main": "index.js",
+  "main": "dist/index.js",
+  "types": "dist/types.d.ts",
   "scripts": {
     "test": "jest",
     "dev-test": "jest --watch",
